rosetta/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
apbs=1.5=h14c3975_3
biopython=1.78=py38h7f8727e_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
collada2gltf=2.1.4=h6bb024c_0
cryptography=36.0.0=py38h9ce1e76_0
cudatoolkit=11.1.1=h6406543_8
curl=7.80.0=h7f8727e_0
dbus=1.13.18=hb2f20db_0
expat=2.4.4=h295c915_0
fftw=3.3.9=h27cfd23_1
fontconfig=2.13.1=h6c09931_0
freemol=1.158=py_2
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glew=2.0.0=0
glib=2.69.1=h4ff587b_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.6.0=py38ha0f2276_0
hdf4=4.2.13=h3ca952b_2
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9d=h7f8727e_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libcurl=7.80.0=h0b77cf5_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglu=9.0.0=hf484d3e_1
libgomp=11.2.0=h1d223b6_14
libholoplaycore=0.1.0_rc4=1
libnetcdf=4.8.1=h42ceab0_1
libnghttp2=1.46.0=hce63b2e_0
libopenblas=0.3.17=hf726d26_1
libpng=1.6.37=hbc83047_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
libzip=1.5.1=h8d318fa_1003
lz4-c=1.9.3=h295c915_1
mengine=1=h14c3975_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpeg_encode=1=h14c3975_1
mtz2ccp4_px=1.0=h9ac9557_3
ncurses=6.3=h7f8727e_2
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
ocl-icd=2.3.1=h7f98852_0
ocl-icd-system=1.0.0=1
openmm=7.6.0=py38h7c39f5a_0
openssl=1.1.1n=h7f8727e_0
pandas=1.1.3=py38he6710b0_0
pcre=8.45=h295c915_0
pdb2pqr=2.1.2+pymol=py_0
pdbfixer=1.8.1=pyh6c4a22f_0
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
pmw=2.0.1+3=py_3
pycollada=0.7.1+bdf414c7=py_1
pycparser=2.21=pyhd3eb1b0_0
pykerberos=1.2.1=py38h7f8727e_3
pymol=2.4.1=py38h4463551_0
pymol-bundle=2.4.1=0
pymol-web-examples=2.4=1
pyopenssl=22.0.0=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytz=2020.1=py_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
rigimol=1.3=2
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38h295c915_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.0=hc218d9a_0
tk=8.6.11=h1ccaba5_0
urllib3=1.26.8=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
